fix: kube pg cluster recover from remote backups (#377)
I've gotten a chance to test out recovering from a database cluster that
lives in another AWS account and found some issues with the current
implementation of the stack that I had to patch in order to get the
recovery to run.

1. Set iam policy for reading from remote bucket
2. Correcting a targetImmediate problem when creating the manifest for
those properties that happens if you try to apply now

```
Planning failed. OpenTofu encountered an error while generating this plan.

╷
│ Error: Manifest configuration incompatible with resource schema
│
│   with module.database.kubernetes_manifest.postgres_cluster,
│   on ../../../../../../../oss/panfactum/stack/packages/infrastructure/kube_pg_cluster/main.tf line 267, in resource "kubernetes_manifest" "postgres_cluster":
│  267: resource "kubernetes_manifest" "postgres_cluster" {
│
│ Detailed descriptions of errors will follow below.
╵
```


Tested in a live cluster by pointing to our production backups.

Documentation reference
https://cloudnative-pg.io/documentation/1.20/backup/

Closes https://github.com/Panfactum/stack/issues/378